feat(market,brapi): add currency/treasury sync, fundamentals, macro/FII endpoints, and market redesign - Extend BrapiPriceUpdateJobService to sync forex pairs (currency/available   + currency/historical backfill) and Tesouro Direto (treasury/list + indicators   + indicators/history) into MarketAsset with types Moeda and TesouroDireto - Add ProcessFundamentalsBatchAsync: daily best-effort upsert of MarketAssetFundamentals   (DY, P/L, P/VP, ROE, marketCap, revenue, net income) via quote?modules= — isolated   from price batch so a modules failure never blocks price update - Add MarketAssetFundamentals entity, EF mapping, and AddDbIndexes migration; add   fundamentals-based sort params (dy_desc, marketcap_desc, pl_asc, pvp_asc, roe_desc)   to GET /api/market - Add GET /api/market/{ticker}/fii (FiiIndicatorsDto) and GET /api/market/macro   (MacroIndicatorDto[]) endpoints with 6h memory cache - Add DB indexes: User.Email unique, User.PasswordResetToken partial, RefreshToken.Token   unique, Tag(UserId,Name) unique, Investment xmin row-version for optimistic concurrency - Add RefreshTokenCleanupJobService + hosted service to purge expired tokens - Redesign /market into a dashboard: MarketIndicatorsStrip (IBOV/IFIX + forex + macro),   RankingCard with fundamentals sorts, MarketAssetRow as Link; replace separate search   components with GlobalSearch "Mercado" section - Add /market/[ticker] (MarketAssetPage with FundamentalsPanel / FiiPanel) and   /market/ranking routes; remove MarketSearchBar and MarketSearchResults (now dead code)

+++ apps/api/FinanceControl.Shared/Dtos/Response/Analytics/FinancialMilestoneDto.cs
@@ -10,7 +10,7 @@ namespace FinanceControl.Shared.Dtos.Response.Analytics
     {
         public string Label { get; set; } = "";
         public DateOnly Date { get; set; }
-        public int NetWorthAtDate { get; set; }
+        public long NetWorthAtDate { get; set; }
         public string Type { get; set; } = "";
     }
 
@@ -18,6 +18,6 @@ namespace FinanceControl.Shared.Dtos.Response.Analytics
     {
         public int Year { get; set; }
         public int Month { get; set; }
-        public int NetWorth { get; set; }
+        public long NetWorth { get; set; }
     }
 }
